Fixed restock issue #389

+++ src/game/CWorld.cpp
@@ -195,7 +195,7 @@ void ReportGarbageCollection(CObjBase * pObj, int iResultCode)
 	DEBUG_ERR(("GC: Deleted UID=0%" PRIx32 ", Defname='%s', Name='%s'. Invalid code=0x%x (%s).\n",
 		(dword)pObj->GetUID(), pObj->Base_GetDef()->GetResourceName(), pObj->GetName(), iResultCode, GetReasonForGarbageCode(iResultCode)));
 
-	if ( (pObj->_iCreatedResScriptIdx != (size_t)-1) && (pObj->_iCreatedResScriptLine != -1) )
+	if ( (pObj->_iCreatedResScriptIdx != -1) && (pObj->_iCreatedResScriptLine != -1) )
 	{
 		// Object was created via NEWITEM or NEWNPC in scripts, tell me where
 		CResourceScript* pResFile = g_Cfg.GetResourceFile((size_t)(pObj->_iCreatedResScriptIdx));
